Extract functions from buttons capabilities
Related to #44
