+++ src/hubitat-capabilities/functions/which-button-held.function.ts
@@ -3,27 +3,31 @@
  * @module HubitatCapabilities
  */
 
-import { HubitatDevice } from '../hubitat-device-events/hubitat-device';
-import { getDevice } from './capabilities.helpers';
+import { HubitatDevice } from '../../hubitat-device-events/hubitat-device';
+import { getDevice } from '../capabilities.helpers';
 
 /**
  * Returns a number of the button that was held.
+ *
+ * Capabilities:
+ * - HoldableButton
  *
  * @param device A target device.
  * @returns A number of the button that was held.
- * @category HoldableButton capability
  */
-export function whichButtonIsHeld(device: HubitatDevice): number;
+export function whichButtonHeld(device: HubitatDevice): number;
 
 /**
  * Returns a number of the button that was held.
  *
+ * Capabilities:
+ * - HoldableButton
+ *
  * @param deviceId An ID of the target device.
  * @returns A number of the button that was held.
- * @category HoldableButton capability
  */
-export function whichButtonIsHeld(deviceId: number): number;
+export function whichButtonHeld(deviceId: number): number;
 
-export function whichButtonIsHeld(deviceOrId: HubitatDevice | number): number {
+export function whichButtonHeld(deviceOrId: HubitatDevice | number): number {
   return getDevice(deviceOrId).getAttributeAsInt('held');
 }
